Main page is generating somewhat of a list on home
The list can be generated through the data within the database, and the utilization of jinja will make all the spaces necessary and so far it will make 3 sections. The next task to make a post request that takes the key value of the table for books will be done tonight.